Move (SE) appendix formatting from .cls to .tex
Previously, this was controlled from within the class. "Appendix" would
prefix every appendix letter, even subsections. So we would get things
like "Appendix A," which is fine, and "Appendix A.1," which doesn't
look as nice.

Thus, the `appendix' package is used in the SE .tex template. This
allows greater control over the appendix formatting.

There may be a way of achieving this without using the `appendix'
package, but it probably isn't as simple.